# src/server/openapi-ai.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Task Priority App API",
    "version": "1.0.0",
    "description": "AI-optimized task management API with ICE scoring (Impact, Confidence, Ease), 4D decision framework (Do, Delegate, Delay, Delete), time tracking, and AI recommendations. Use this API to manage tasks, track time, get productivity analytics, and manage programmatic access via API keys.",
    "contact": {
      "name": "API Support",
      "url": "https://github.com/yourusername/task-priority-app"
    },
    "x-ai-friendly": true,
    "x-supports-function-calling": true
  },
  "servers": [
    {
      "url": "http://localhost:8787",
      "description": "Development server"
    },
    {
      "url": "https://your-app.workers.dev",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "description": "API key authentication. Use format: 'Bearer task_live_XXXXXXXX...' in Authorization header"
      }
    },
    "schemas": {
      "Task": {
        "type": "object",
        "description": "A task with ICE scoring for prioritization",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique task identifier"
          },
          "name": {
            "type": "string",
            "description": "Task title or name",
            "example": "Complete project documentation"
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "description": "Additional notes or description in plain text or markdown"
          },
          "impact": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10,
            "description": "Impact score (1-10): How much value this task provides when completed",
            "example": 8
          },
          "confidence": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10,
            "description": "Confidence score (1-10): How confident you are in your ability to complete this",
            "example": 7
          },
          "ease": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10,
            "description": "Ease score (1-10): How easy/quick this task is to complete",
            "example": 6
          },
          "type": {
            "type": "string",
            "enum": ["revenue", "growth", "operations", "strategic", "personal"],
            "description": "Task category: revenue (income-generating), growth (learning/skills), operations (day-to-day), strategic (long-term planning), personal (personal tasks)"
          },
          "timeBlock": {
            "type": "string",
            "enum": ["deep", "collaborative", "quick", "systematic"],
            "description": "When to do this: deep (focused deep work), collaborative (with others), quick (small tasks), systematic (routine processes)"
          },
          "estimatedTime": {
            "type": "integer",
            "description": "Estimated duration in minutes",
            "example": 120
          },
          "decision": {
            "type": "string",
            "enum": ["do", "delegate", "delay", "delete"],
            "description": "4D Framework decision: do (prioritize now), delegate (assign to others), delay (postpone), delete (eliminate)"
          },
          "status": {
            "type": "string",
            "enum": ["active", "completed", "archived"],
            "default": "active",
            "description": "Task status"
          },
          "actualTime": {
            "type": "integer",
            "default": 0,
            "description": "Total time spent on task in minutes (tracked via focus sessions)"
          },
          "isInFocus": {
            "type": "boolean",
            "description": "Whether task currently has an active time tracking session"
          },
          "deadline": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Hard deadline for task completion (ISO 8601 format)",
            "example": "2025-12-31T23:59:59Z"
          },
          "scheduledFor": {
            "type": "string",
            "enum": ["today", "this-week", "this-month", "someday"],
            "nullable": true,
            "description": "Time horizon for task completion"
          },
          "recurringPattern": {
            "type": "string",
            "enum": ["daily", "weekly", "monthly"],
            "nullable": true,
            "description": "Recurrence pattern for repeating tasks"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When task was created"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "When task was completed"
          }
        },
        "required": ["name", "impact", "confidence", "ease", "type", "timeBlock", "estimatedTime", "decision"]
      },
      "CreateTaskRequest": {
        "type": "object",
        "description": "Request body for creating a new task",
        "required": ["name", "impact", "confidence", "ease", "type", "timeBlock", "estimatedTime", "decision"],
        "properties": {
          "name": {
            "type": "string",
            "description": "Task title",
            "example": "Write blog post about productivity"
          },
          "notes": {
            "type": "string",
            "description": "Optional detailed notes",
            "example": "Cover ICE scoring methodology and examples"
          },
          "impact": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10,
            "example": 8
          },
          "confidence": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10,
            "example": 7
          },
          "ease": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10,
            "example": 6
          },
          "type": {
            "type": "string",
            "enum": ["revenue", "growth", "operations", "strategic", "personal"],
            "example": "growth"
          },
          "timeBlock": {
            "type": "string",
            "enum": ["deep", "collaborative", "quick", "systematic"],
            "example": "deep"
          },
          "estimatedTime": {
            "type": "integer",
            "description": "Estimated minutes to complete",
            "example": 120
          },
          "decision": {
            "type": "string",
            "enum": ["do", "delegate", "delay", "delete"],
            "example": "do"
          },
          "scheduledFor": {
            "type": "string",
            "enum": ["today", "this-week", "this-month", "someday"],
            "example": "this-week"
          },
          "deadline": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": "2025-12-31T23:59:59Z"
          }
        },
        "example": {
          "name": "Write blog post about productivity",
          "impact": 8,
          "confidence": 7,
          "ease": 6,
          "type": "growth",
          "timeBlock": "deep",
          "estimatedTime": 120,
          "decision": "do",
          "scheduledFor": "this-week"
        }
      },
      "ApiKey": {
        "type": "object",
        "description": "API key for programmatic access",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Descriptive name for this key",
            "example": "Mobile App Integration"
          },
          "prefix": {
            "type": "string",
            "description": "First 8 characters for identification",
            "example": "task_abc1"
          },
          "key": {
            "type": "string",
            "description": "Full API key (only shown once on creation)",
            "example": "task_live_abc123def456..."
          },
          "lastUsedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "requestCount": {
            "type": "integer",
            "description": "Total requests made with this key"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "details": {
            "type": "object",
            "description": "Additional error details"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/tasks": {
      "get": {
        "operationId": "listTasks",
        "summary": "List all tasks",
        "description": "Retrieve all tasks for the authenticated user. Optionally filter by status (active, completed, archived). Returns tasks with ICE scores, time tracking data, and decision framework classification.",
        "tags": ["Tasks"],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter tasks by status",
            "schema": {
              "type": "string",
              "enum": ["active", "completed", "archived"],
              "example": "active"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of tasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tasks": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Task"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createTask",
        "summary": "Create a new task",
        "description": "Create a new task with ICE scoring (Impact, Confidence, Ease). The task will be prioritized using AI algorithms based on these scores and the 4D decision framework (Do, Delegate, Delay, Delete).",
        "tags": ["Tasks"],
        "requestBody": {
          "required": true,
          "description": "Task details including required ICE scores and categorization",
          "content": {
            "application/json": {
                "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "task": {
                      "$ref": "#/components/schemas/Task"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error - Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Validation error",
                  "details": {
                    "impact": "Must be between 1 and 10"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/stats": {
      "get": {
        "operationId": "getStatistics",
        "summary": "Get productivity statistics",
        "description": "Retrieve comprehensive productivity statistics including task counts, time spent, completion rates, and AI-generated recommendations based on 8 different prioritization algorithms (simple, weighted, roi, eisenhower, skill, energy, strategic, hybrid).",
        "tags": ["Analytics"],
        "parameters": [
          {
            "name": "method",
            "in": "query",
            "description": "AI recommendation algorithm to use for task prioritization",
            "schema": {
              "type": "string",
              "enum": ["simple", "weighted", "roi", "eisenhower", "skill", "energy", "strategic", "hybrid"],
              "default": "hybrid",
              "example": "hybrid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Statistics and AI recommendations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "totalTasks": {
                      "type": "integer"
                    },
                    "completedTasks": {
                      "type": "integer"
                    },
                    "activeTasks": {
                      "type": "integer"
                    },
                    "totalTimeSpent": {
                      "type": "integer",
                      "description": "Total minutes tracked across all tasks"
                    },
                    "tasksByDecision": {
                      "type": "object",
                      "properties": {
                        "do": {
                          "type": "integer"
                        },
                        "delegate": {
                          "type": "integer"
                        },
                        "delay": {
                          "type": "integer"
                        },
                        "delete": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/api-keys": {
      "get": {
        "operationId": "listApiKeys",
        "summary": "List API keys",
        "description": "List all API keys created by the authenticated user, including usage statistics (last used, request count) and expiration dates.",
        "tags": ["API Keys"],
        "responses": {
          "200": {
            "description": "List of API keys",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "apiKeys": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ApiKey"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createApiKey",
        "summary": "Create API key",
        "description": "Create a new API key for programmatic access. IMPORTANT: The full key is only returned once at creation time - save it securely. The key format is 'task_live_XXXXXXXX...' and must be sent as 'Authorization: Bearer <key>' in subsequent requests.",
        "tags": ["API Keys"],
        "requestBody": {
          "required": true,
          "description": "API key configuration",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100,
                    "description": "Descriptive name for this key (e.g., 'Mobile App', 'Zapier Integration')",
                    "example": "Mobile App"
                  },
                  "expiresInDays": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 365,
                    "description": "Optional: Days until key expires (omit for no expiration)",
                    "example": 90
                  }
                },
                "example": {
                  "name": "Mobile App",
                  "expiresInDays": 90
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API key created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "apiKey": {
                      "$ref": "#/components/schemas/ApiKey"
                    },
                    "message": {
                      "type": "string",
                      "example": "API key created successfully. Save this key securely - it won't be shown again."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/api-keys/{id}": {
      "delete": {
        "operationId": "deleteApiKey",
        "summary": "Delete API key",
        "description": "Delete an API key. This immediately revokes access for the key. This action cannot be undone.",
        "tags": ["API Keys"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "API key ID (UUID)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API key deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "API key deleted successfully"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
...
}